+++ game-hollowknightsilksong/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
-## [0.1.0] - 2025-06-27
+## [0.1.1] - 2025-09-04
+
+- Added Assembly-CSharp.dll mod installer
+- Added buttons to open the Data and save folders (folder icons on Mods page toolbar)
+
+## [0.1.0] - 2025-09-04
 
 - Initial release